requirements.txt
streamlit>=1.30.0
openai>=1.0.0
langchain>=0.1.20
langchain-community>=0.0.20
python-dotenv>=1.0.1
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
tiktoken>=0.6.0